requirements.txt
django
currencyconverter
pyalex
beautifulsoup4
nameparser
playwright
asyncio
Unidecode
django-allauth
django-data-browser
django-debug-toolbar
habanero
more_itertools
xmltodict
psycopg[binary]
django_extensions
django-redis
pybliometrics
django-ajax-datatable
rispy
numpy
nltk
jellyfish
httpx
pandas
pillow
pymongo
python-dotenv